Refactor - Separate into different files
